run tests in many browsers (travis + saucelabs)

+++ test/spec/index/test-cases/deleteUserKey.js
@@ -1,12 +1,16 @@
-module.exports = {
-  object: 'index',
-  methodName: 'deleteUserKey',
-  testName: 'index.deleteUserKey(key, cb)',
-  callArguments: ['lk9089lk'],
-  expectedRequest: {
-    method: 'DELETE',
-    URL: {
-      pathname: '/1/indexes/%s/keys/lk9089lk'
+var fauxJax = require('faux-jax');
+
+if (fauxJax.support.cors) {
+  module.exports = {
+    object: 'index',
+    methodName: 'deleteUserKey',
+    testName: 'index.deleteUserKey(key, cb)',
+    callArguments: ['lk9089lk'],
+    expectedRequest: {
+      method: 'DELETE',
+      URL: {
+        pathname: '/1/indexes/%s/keys/lk9089lk'
+      }
     }
-  }
-};
+  };
+}
